+++ CHANGELOG.md
@@ -4,6 +4,7 @@ CHANGELOG
 0.3
 ---
 
+ * Add support for `extension: false` flag in `extra.ai-mate` composer.json configuration to exclude packages from being discovered as extensions
  * Add `ToolsListCommand` to list all available tools
 
 0.2